pyproject.toml
[tool.poetry]
name = "minicms"
version = "0.1.0"
description = ""
authors = ["Joackim Balouka <jobalouka@gmail.com>"]

[tool.poetry.dependencies]
python = "3.7.6"
django = "^3.0.7"
psycopg2 = "^2.8.5"
ipython = "^7.15.0"
path = "^14.0.1"
gunicorn = "^20.0.4"
dj-database-url = "^0.5.0"
python-dotenv = "^0.13.0"
whitenoise = "^5.1.0"
django-ckeditor = "^5.9.0"
djangorestframework = "^3.11.0"
django-cors-headers = "^3.4.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
